Fix perf issues in bus, PPU
The bus now uses a ~66k array with a 1:1 correspondence between address locations and device mappings. This seems a little excessive at first glance, but memory is dirt cheap and the earlier mapper was burning a lot of cycles looping through each device to find the one mapped to that memory location.